refactor: separated the core logic from route file
